+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master (unreleased)
 
 * Update code to work with rubocop >= 0.19 ([@geniou][])
-
+* Split `UnitSpecNaming` into `RSpecDescribedClass`, `RSpecDescribedMethod` and
+  `RSpecFileName` and enabled them all by default ([@geniou][])
 
 [@geniou]: https://github.com/geniou